# openapi/api.yaml
...
openapi: 3.0.3
info:
  title: SurenSaaS API
  version: 1.0.0
  description: API multi-PME pour construction, nettoyage et diagnostic énergétique
servers:
  - url: https://back-xxx.run.app/api/v1
    description: Production
paths:
...
  /invoices:
    get:
      summary: Lister les factures
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [brouillon, en_attente_validation, validee, rejetee, en_traitement_comptable, archivee]
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Liste des factures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'

    post:
      summary: Créer une facture
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        201:
          description: Facture créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceV2'
  /invoices/{invoice_id}:
    get:
      summary: Détails d'une facture
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Détails de la facture
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceV2'
        404:
          description: Facture non trouvée

    put:
      summary: Modifier une facture (brouillon uniquement)
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        200:
          description: Facture modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceV2'
        400:
          description: Facture non modifiable (pas en brouillon)

    delete:
      summary: Supprimer une facture (brouillon uniquement)
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Facture supprimée
        400:
          description: Facture non supprimable (pas en brouillon)
  /invoices/{invoice_id}/validate:
    post:
      summary: Valider ou rejeter une facture
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum: [validate, reject]
      responses:
        200:
          description: Facture mise à jour
        400:
          description: Action invalide ou facture non en attente

  # ============================================
  # TELEGRAM BOT
  # ============================================
...
  /{org}/companies/{company_id}/invoices:
    get:
      summary: Lister les factures
      security:
        - bearerAuth: []
      description: Nécessite capability construction:facturation:read ou rôle admin
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [brouillon, en_attente_validation, validee, rejetee, en_traitement_comptable, archivee]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Liste des factures
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  total:
                    type: integer
        403:
          description: Capability manquante

    post:
      summary: Créer une facture (manuel ou bot)
      security:
        - bearerAuth: []
      description: Nécessite capability construction:facturation:write ou rôle admin
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        201:
          description: Facture créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        403:
          description: Capability manquante
  /{org}/companies/{company_id}/invoices/{invoice_id}:
    get:
      summary: Détails d'une facture
      security:
        - bearerAuth: []
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Détails de la facture
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        404:
          description: Facture non trouvée

    put:
      summary: Modifier une facture
      security:
        - bearerAuth: []
      description: Nécessite capability construction:facturation:write ou rôle admin. Ne fonctionne que sur les factures en brouillon.
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvoiceRequest'
      responses:
        200:
          description: Facture modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        403:
          description: Non autorisé ou facture non modifiable

    delete:
      summary: Supprimer une facture
      security:
        - bearerAuth: []
      description: Nécessite rôle admin
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Facture supprimée
        403:
          description: Non autorisé
  /{org}/companies/{company_id}/invoices/{invoice_id}/validate:
    post:
      summary: Valider ou rejeter une facture
      security:
        - bearerAuth: []
      description: Nécessite capability construction:facturation:validate ou rôle admin
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateInvoiceRequest'
      responses:
        200:
          description: Facture mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        403:
          description: Capability manquante
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoice_number:
          type: string
          nullable: true
        supplier_name:
          type: string
        amount_ttc:
          type: number
          format: decimal
        invoice_date:
          type: string
          format: date
        status:
          type: string
          enum: [brouillon, en_attente_validation, validee, rejetee, en_traitement_comptable, archivee]
        created_at:
          type: string
          format: date-time
        created_by_telegram:
          type: boolean
    InvoiceDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/Invoice'
        - type: object
          properties:
            supplier_address:
              type: string
            supplier_siret:
              type: string
            amount_ht:
              type: number
              format: decimal
            vat_amount:
              type: number
              format: decimal
            vat_rate:
              type: number
            due_date:
              type: string
              format: date
            description:
              type: string
            items:
              type: array
              items:
                type: object
            original_file_url:
              type: string
            thumbnail_url:
              type: string
            validated_by:
              type: string
              format: uuid
            validated_at:
              type: string
              format: date-time
            rejection_reason:
              type: string
            ocr_data:
              type: object
    CreateInvoiceRequest:
      type: object
      required:
        - supplier_name
        - amount_ttc
      properties:
        supplier_name:
          type: string
        supplier_address:
          type: string
        supplier_siret:
          type: string
        amount_ht:
          type: number
          format: decimal
        amount_ttc:
          type: number
          format: decimal
        vat_amount:
          type: number
          format: decimal
        vat_rate:
          type: number
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        description:
          type: string
        invoice_number:
          type: string
        original_file_url:
          type: string
        status:
          type: string
          enum: [brouillon, en_attente_validation]
          default: brouillon
    UpdateInvoiceRequest:
      type: object
      properties:
        supplier_name:
          type: string
        supplier_address:
          type: string
        supplier_siret:
          type: string
        amount_ht:
          type: number
          format: decimal
        amount_ttc:
          type: number
          format: decimal
        vat_amount:
          type: number
          format: decimal
        vat_rate:
          type: number
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        description:
          type: string
        invoice_number:
          type: string
    ValidateInvoiceRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [validate, reject]
        reason:
          type: string
          description: Raison du rejet (si action=reject)
...